// src/app/interfaces/player.interface.ts
export interface Player{
    $key?: string;
    name: string;
    lastName: string;
    position: SquadNumber;
    weight: number;
    height: number;
    nationality: Countries;
    leftFooted: boolean   // es izquierdo?

}

export enum SquadNumber {
    goalkeeper = 1,
    rigthBack = 4,
    sweeper = 2,
    stopper = 6,
    lefBack = 3,
    RightMidFielder = 8,
    centralDefensiveMidFielder = 5,
    lefMidFielder = 10,
    righWinger = 7,
    centerForward = 9,
    lefWinger = 11

}